Bias Correction for Simulated Climate Data

Applies statistical bias correction to simulated climate variables using observed reference data.

bias_correction(
  simulated_data,
  observed_data,
  variables = c("Tmin", "Tmax", "Avg.Temp", "Rainfall", "RH", "WindSpeed",
    "Solar_Radiation", "ET0"),
  method = "mean_scaling",
  digits = 2,
  return_factors = TRUE
)

Arguments

simulated_data

Simulated climate data frame.

observed_data

Observed/reference climate data frame.

variables

Character vector of variables to correct.

method

Bias-correction method. Options are: "mean_scaling", "additive", "multiplicative".

digits

Number of decimal places.

return_factors

Logical. If TRUE, returns correction factors and bias statistics.

Value

A list containing:

corrected_data

Bias-corrected climate dataset.

correction_factors

Bias-correction factors and bias statistics applied to each variable.

Details

Supported correction methods include:

Mean scaling

Additive correction

Multiplicative correction

Physical constraints are automatically enforced after correction to ensure climatological realism.

Examples

if (FALSE) { # \dontrun{

bc <- bias_correction(
  simulated_data = cd,
  observed_data = obs_data,
  variables = c("Rainfall", "Avg.Temp")
)

head(bc$corrected_data)
bc$correction_factors

} # }