Apply Copula Dependence Structure to Climate Variables

Introduces multivariate dependence among simulated climate variables using Gaussian or t copulas.

copula_dependence(
  climate_data,
  variables = c("Tmin", "Tmax", "Avg.Temp", "Rainfall", "RH", "WindSpeed",
    "Solar_Radiation", "ET0"),
  copula_type = "gaussian",
  df = 4,
  seed = 123,
  digits = 2
)

Arguments

climate_data

Climate data frame.

variables

Variables used in dependence modeling.

copula_type

Type of copula. Options are: "gaussian" or "t".

df

Degrees of freedom for t copula. Default is 4.

seed

Random seed for reproducibility.

digits

Number of decimal places.

Value

A list containing:

adjusted_data

Climate dataset with copula-adjusted dependence structure.

correlation_matrix

Empirical correlation matrix used in copula fitting.

copula_type

Copula family used.

Details

This function improves realism by preserving correlations and inter-variable dependency structures commonly observed in climate systems.

Supported variables include:

Tmin

Tmax

Avg.Temp

Rainfall

RH

WindSpeed

Solar_Radiation

ET0

Examples

if (FALSE) { # \dontrun{

cp <- copula_dependence(cd)

head(cp$adjusted_data)

} # }